Renamed enum from State (cherry picked to rename ReduxState)
## Problem
The original design of Breadwallet (Loafwallet) used a SimpleRedux architecture which maanged the State in an unsual (yet effective) way.

Within that design ‘state’ was managed by an enum named “State”

This was fine until SwiftUI and Combine where there is now a name collision since @State is referring to the SwfitUI attribute.

## Approach
While redsiging the app would be ideal, the reality is this needs to be fixed as soon as possible because Litewallet is now using SwiftUI.

Steps to fix the problem:
- [ ] Renamed variable from `State` to `ReduxState`
- [ ] Replaced all cases where `State` and `state` where used with `ReduxState` and `reduxState`

# Conflicts:
#	loafwallet/DonationSetupCell.swift
#	loafwallet/DynamicDonationViewController.swift
#	loafwallet/src/ViewControllers/RootModals/SendViewController.swift